requirements.txt
streamlit==1.50.0
pandas==2.2.2
plotly==5.22.0
matplotlib==3.8.4
altair==5.5.0
PyPDF2==3.0.1
numpy==1.26.4
pillow==10.3.0
